+++ inst/doc/FKFSP.R
@@ -5,24 +5,24 @@ knitr::opts_chunk$set(
 )
 
 ## ----setup--------------------------------------------------------------------
+##The packages 'FKF', 'stats' and 'NFCP' are required for this Vignette:
 library(FKF.SP)
-##The package 'FKF' is required for this Vignette:
-# install.packages("FKF")
 library(FKF)
+library(stats)
+library(NFCP)
 
 ## -----------------------------------------------------------------------------
+# Set constants:
 ## Length of series
 n <- 10000
 
 ## AR parameters
-ar1 <- 0.6
-ar2 <- 0.2
-ma1 <- -0.2
-sigma <- sqrt(0.2)
+AR <- c(ar1 = 0.6, ar2 = 0.2, ma1 = -0.2, sigma = sqrt(0.2))
 
-## -----------------------------------------------------------------------------
-a <- stats::arima.sim(model = list(ar = c(ar1, ar2), ma = ma1), n = n,
-            innov = rnorm(n) * sigma)
+# Generate observations:
+set.seed(1)
+a <- stats::arima.sim(model = list(ar = AR[c("ar1", "ar2")], ma = AR["ma1"]), n = n,
+            innov = rnorm(n) * AR["sigma"])
 
 ## -----------------------------------------------------------------------------
 arma21ss <- function(ar1, ar2, ma1, sigma) {
@@ -37,20 +37,19 @@ a0 <- c(0, 0)
 ## Diffuse assumption
 P0 <- matrix(1e6, nrow = 2, ncol = 2)
 return(list(a0 = a0, P0 = P0, ct = ct, dt = dt, Zt = Zt, Tt = Tt, GGt = GGt,
-            HHt = HHt))
-            }
+            HHt = HHt))}
 
 ## -----------------------------------------------------------------------------
 # The objective function passed to 'optim'
-objective <- function(theta, yt, SP = F) {
+objective <- function(theta, yt, SP) {
 param <- arma21ss(theta["ar1"], theta["ar2"], theta["ma1"], theta["sigma"])
-# Kalman Filtering through the fkf.SP function:
+# Kalman filtering through the 'fkf.SP' function:
 if(SP){
  ans <- - fkf.SP(a0 = param$a0, P0 = param$P0, dt = param$dt, ct = param$ct, 
                Tt = param$Tt, Zt = param$Zt, HHt = param$HHt, GGt = param$GGt, 
                yt = yt)
  }
-# Kalman Filtering through the fkf function:
+# Kalman filtering through the 'fkf' function:
  else{
  ans <- - fkf(a0 = param$a0, P0 = param$P0, dt = param$dt, ct = param$ct, Tt = param$Tt,
             Zt = param$Zt, HHt = param$HHt, GGt = param$GGt, yt = yt)$logLik
@@ -58,19 +57,23 @@ if(SP){
  }
  return(ans)
 }
-##Optim minimizes functions, so the negative is returned
+##Optim minimizes functions by default, so the negative is returned
 
 
 ## -----------------------------------------------------------------------------
+#This test estimates parameters through 'optim'.
+#Please run the complete chunk for a fair comparison:
+
+#Initial values:
 theta <- c(ar = c(0, 0), ma1 = 0, sigma = 1)
 
-###FKF Package:
+###MLE through the 'fkf' function:
 start <- Sys.time()
 set.seed(1)
 FKF_estimation <- optim(theta, objective, yt = rbind(a), hessian = TRUE, SP = F)
-FKF_runtime = Sys.time() - start
+FKF_runtime <- Sys.time() - start
 
-###fkf.SP Package:
+###MLE through the 'fkf.SP' function:
 start <- Sys.time()
 set.seed(1)
 FKF.SP_estimation <- optim(theta, objective, yt = rbind(a), hessian = TRUE, SP = T)
@@ -78,15 +81,13 @@ FKF.SP_runtime <- Sys.time() - start
 
 
 ## -----------------------------------------------------------------------------
-print(rbind(FKF.SP_estimation$par, FKF_estimation$par))
+print(rbind(FKF.SP = FKF.SP_estimation$par, FKF = FKF_estimation$par))
 
 ## -----------------------------------------------------------------------------
 print(c(FKF.SP = FKF.SP_estimation$counts[1], FKF = FKF_estimation$counts[1]))
 
 ## -----------------------------------------------------------------------------
-
 print(c(FKF.SP = FKF.SP_runtime, FKF = FKF_runtime))
-
 
 ## -----------------------------------------------------------------------------
 ## Transition equation:
@@ -111,15 +112,15 @@ Nile_MLE <- function(yt, SP){
 ##Unknown parameters initial estimates:
 GGt <- HHt <- var(yt, na.rm = TRUE) * .5
 set.seed(1)
-#fkf.SP function:
+# Kalman filtering through the 'fkf.SP' function:
 if(SP){
-  return(suppressWarnings(optim(c(HHt = HHt, GGt = GGt),
+  return(optim(c(HHt = HHt, GGt = GGt),
         fn = function(par, ...)
              -fkf.SP(HHt = matrix(par[1]), GGt = matrix(par[2]), ...),
              yt = rbind(yt), a0 = a0, P0 = P0, dt = dt, ct = ct,
-             Zt = Zt, Tt = Tt)))
+             Zt = Zt, Tt = Tt))
 } else {
-#fkf function:
+# Kalman filtering through the 'fkf' function:
   return(optim(c(HHt = HHt, GGt = GGt),
         fn = function(par, ...)
              -fkf(HHt = matrix(par[1]), GGt = matrix(par[2]), ...)$logLik,
@@ -128,8 +129,8 @@ if(SP){
 }}
 
 ## -----------------------------------------------------------------------------
-fkf.SP_MLE_complete = Nile_MLE(y_complete, SP = T)
-fkf_MLE_complete = Nile_MLE(y_complete, SP = F)
+fkf.SP_MLE_complete <- Nile_MLE(y_complete, SP = T)
+fkf_MLE_complete <- Nile_MLE(y_complete, SP = F)
 
 ## -----------------------------------------------------------------------------
 print(fkf.SP_MLE_complete[1:3])
@@ -138,8 +139,8 @@ print(fkf.SP_MLE_complete[1:3])
 print(fkf_MLE_complete[1:3])
 
 ## -----------------------------------------------------------------------------
-fkf.SP_MLE_incomplete = Nile_MLE(y_incomplete, SP = T)
-fkf_MLE_incomplete = Nile_MLE(y_incomplete, SP = F)
+fkf.SP_MLE_incomplete <- Nile_MLE(y_incomplete, SP = T)
+fkf_MLE_incomplete <- Nile_MLE(y_incomplete, SP = F)
 
 ## -----------------------------------------------------------------------------
 print(fkf.SP_MLE_incomplete[1:3])
@@ -149,7 +150,7 @@ print(fkf_MLE_incomplete[1:3])
 
 ## -----------------------------------------------------------------------------
 #Number of NA values:
-NA_values = length(which(is.na(y_incomplete)))
+NA_values <- length(which(is.na(y_incomplete)))
 
 print( 0.5 * NA_values * log(2 * pi))
 
@@ -157,22 +158,26 @@ print( 0.5 * NA_values * log(2 * pi))
 #This test uses estimated parameters of complete data. 
 #Please run the complete chunk for a fair comparison:
 
+#'fkf' 
 set.seed(1)
-start = Sys.time()
+start <- Sys.time()
 for(i in 1:1e4) fkf(a0, P0, dt, ct, Tt, Zt, HHt = matrix(fkf_MLE_complete$par[1]),
                     GGt = matrix(fkf_MLE_complete$par[2]), yt = rbind(y_complete))
-FKF_runtime = Sys.time() - start
+FKF_runtime <- Sys.time() - start
 
-start = Sys.time()
+#'fkf.SP'
 set.seed(1)
+start = Sys.time()
 for(i in 1:1e4) fkf.SP(a0, P0, dt, ct, Tt, Zt, HHt = matrix(fkf.SP_MLE_complete$par[1]),
                        GGt = matrix(fkf.SP_MLE_complete$par[2]), yt = rbind(y_complete))
-fkf.SP_runtime = Sys.time() - start
+fkf.SP_runtime <- Sys.time() - start
 
 print(c(FKF.SP = fkf.SP_runtime, FKF = FKF_runtime))
 
 
 ## -----------------------------------------------------------------------------
+#This test estimates parameters 10 times through 'optim'.
+#Please run the complete chunk for a fair comparison:
 
 ## Transition equation:
 ## alpha[t+1] = alpha[t] + eta[t], eta[t] ~ N(0, HHt)
@@ -189,6 +194,8 @@ a0 <- y[1]            # Estimation of the first width
 P0 <- matrix(100)     # Variance of 'a0'
 
 ##Time comparison - Estimate parameters 10 times:
+
+###MLE through the 'fkf' function:
 start = Sys.time()
 set.seed(1)
 for(i in 1:10)  fit.fkf <- optim(c(HHt = var(y, na.rm = TRUE) * .5,
@@ -200,25 +207,88 @@ for(i in 1:10)  fit.fkf <- optim(c(HHt = var(y, na.rm = TRUE) * .5,
 
 run.time_FKF = Sys.time() - start
 
-
+###MLE through the 'fkf.SP' function:
 start = Sys.time()
-##When FKF.SP is input poorly specified parameters 
-##(ie. log-likelihood = NA) is output a warning:
 set.seed(1)
-suppressWarnings(
 for(i in 1:10)  fit.fkf.SP <- optim(c(HHt = var(y, na.rm = TRUE) * .5,
                         GGt = var(y, na.rm = TRUE) * .5),
                       fn = function(par, ...)
                         -fkf.SP(HHt = array(par[1],c(1,1,1)), GGt = matrix(par[2]), ...),
                       yt = rbind(y), a0 = a0, P0 = P0, dt = dt, ct = ct,
                       Zt = Zt, Tt = Tt)
-)
 run.time_FKF.SP = Sys.time() - start
 
 print(c(fkf.SP = run.time_FKF.SP, fkf = run.time_FKF))
 
-## Filter tree ring data with estimated parameters using fkf:
+## Filter tree ring data with estimated parameters using 'fkf':
 fkf.obj <- fkf(a0, P0, dt, ct, Tt, Zt, HHt = array(fit.fkf$par[1],c(1,1,1)),
                GGt = array(fit.fkf$par[2],c(1,1,1)), yt = rbind(y))
 
+
+## -----------------------------------------------------------------------------
+
+yt = t(log(NFCP::SS.Oil$Contracts)) # quoted log futures prices
+delta.t <- NFCP::SS.Oil$dt # Discrete time step
+##time to maturity of quoted futures contracts:
+TTM <- t(NFCP::SS.Oil$Contract.Maturities)
+
+a0 <- yt[1,1]     # initial estimate
+P0 <- matrix(100) # Variance of 'a0'
+
+## GBM Function
+gbm.mle <- function(theta, SP){
+
+ct <- theta["alpha.rn"] * TTM
+dt <- (theta["alpha"] - 0.5 * theta["sigma"]^2) * delta.t
+Zt <- matrix(1, nrow(yt))
+HHt <- matrix(theta["sigma"]^2 * delta.t)
+Tt <- matrix(1)
+
+##'fkf.SP' requires a vector of the diagonal elements of the variances of the measurement error 
+if(SP){
+GGt = rep(theta["sigma.epsilon"]^2, nrow(yt))
+} else {
+##'fkf' instead requires a matrix of the elements of the variances of the measurement error 
+GGt = diag(theta["sigma.epsilon"]^2, nrow(yt))
+}
+
+logLik = ifelse(SP,
+                - fkf.SP(a0 = a0, P0 = P0, dt = dt, ct = ct, Tt = Tt, Zt = Zt, HHt = HHt, GGt = GGt, yt = yt),
+                - fkf(a0 = a0, P0 = P0, dt = dt, ct = ct, Tt = Tt, Zt = Zt, HHt = HHt, GGt = GGt, yt = yt)$logLik
+                )
+return(logLik)
+}
+
+
+## -----------------------------------------------------------------------------
+#This test estimates parameters through 'optim'.
+#Please run the complete chunk for a fair comparison:
+
+#Initial estimates
+gbm.par <- c(alpha = 0, alpha.rn = 0.01, sigma = 0.1, sigma.epsilon = 0.05)
+
+###MLE through the 'fkf.SP' function:
+set.seed(1)
+start = Sys.time()
+fkf.SP.gbm = optim(par = gbm.par, fn = gbm.mle, SP = T)
+fkf.SP_runtime <- Sys.time() - start
+
+###MLE through the 'fkf' function:
+set.seed(1)
+start = Sys.time()
+fkf.gbm = optim(par = gbm.par, fn = gbm.mle, SP = F)
+fkf_runtime <- Sys.time() - start
+
+
+## -----------------------------------------------------------------------------
+print(rbind(FKF.SP = - fkf.SP.gbm$value, FKF = - fkf.gbm$value))
+
+## -----------------------------------------------------------------------------
+print(rbind(FKF.SP = fkf.SP.gbm$par, FKF = fkf.gbm$par))
+
+## -----------------------------------------------------------------------------
+print(c(FKF.SP = fkf.SP.gbm$counts[1], FKF = fkf.gbm$counts[1]))
+
+## -----------------------------------------------------------------------------
+print(c(FKF.SP = fkf.SP_runtime, FKF = fkf_runtime))