+++ NEWS.md
@@ -1,3 +1,12 @@
+# FKF.SP 0.1.1
+
+- Minor fixes in error reporting of 'fkf.SP'.
+- Example 4 of the vignette added
+- Now suggests the 'NFCP' package for example 4
+- Minor documentation edits
+- 'fkf.SP' no longer returns a 'warning' when NA's are returned
+
+
 # FKF.SP 0.1.0
 
 - Release of FKF.SP